refactor(io): move `*_exact` and `*_all` to ext
