+++ compio-io/src/read/mod.rs
@@ -1,6 +1,6 @@
 use std::{io::Cursor, rc::Rc, sync::Arc};
 
-use compio_buf::{BufResult, IoBuf, IoBufMut, IoVectoredBufMut};
+use compio_buf::{BufResult, IoBufMut, IoVectoredBufMut};
 
 mod buf;
 mod ext;
@@ -24,65 +24,6 @@ pub trait AsyncRead {
     async fn read_vectored<V: IoVectoredBufMut>(&mut self, buf: V) -> BufResult<usize, V>
     where
         V::Item: IoBufMut;
-
-    /// Read the exact number of bytes required to fill the buf.
-    async fn read_exact<T: IoBufMut>(&mut self, mut buf: T) -> BufResult<usize, T> {
-        let len = buf.buf_capacity() - buf.buf_len();
-        let mut read = 0;
-        while read < len {
-            let BufResult(result, buf_returned) = self.read(buf).await;
-            buf = buf_returned;
-            match result {
-                Ok(0) => {
-                    return BufResult(
-                        Err(std::io::Error::new(
-                            std::io::ErrorKind::UnexpectedEof,
-                            "failed to fill whole buffer",
-                        )),
-                        buf,
-                    );
-                }
-                Ok(n) => {
-                    read += n;
-                    unsafe { buf.set_buf_init(read) };
-                }
-                Err(ref e) if e.kind() == std::io::ErrorKind::Interrupted => {}
-                Err(e) => return BufResult(Err(e), buf),
-            }
-        }
-        BufResult(Ok(read), buf)
-    }
-
-    /// Read the exact number of bytes required to fill the vector buf.
-    async fn read_vectored_exact<T: IoVectoredBufMut>(&mut self, mut buf: T) -> BufResult<usize, T>
-    where
-        T::Item: IoBufMut,
-    {
-        let len = buf
-            .buf_iter_mut()
-            .map(|x| x.buf_capacity() - x.buf_len())
-            .sum();
-        let mut read = 0;
-        while read < len {
-            let BufResult(res, buf_returned) = self.read_vectored(buf).await;
-            buf = buf_returned;
-            match res {
-                Ok(0) => {
-                    return BufResult(
-                        Err(std::io::Error::new(
-                            std::io::ErrorKind::UnexpectedEof,
-                            "failed to fill whole buffer",
-                        )),
-                        buf,
-                    );
-                }
-                Ok(n) => read += n,
-                Err(ref e) if e.kind() == std::io::ErrorKind::Interrupted => {}
-                Err(e) => return BufResult(Err(e), buf),
-            }
-        }
-        BufResult(Ok(read), buf)
-    }
 }
 
 macro_rules! impl_read {
