+++ compio-io/src/write/ext.rs
@@ -1,4 +1,6 @@
-use crate::{AsyncWrite, IoResult};
+use compio_buf::{buf_try, BufResult, IntoInner, IoBuf};
+
+use crate::{AsyncWrite, AsyncWriteAt, IoResult};
 
 macro_rules! write_scalar {
     ($t:ty, $be:ident, $le:ident) => {
@@ -31,6 +33,19 @@ macro_rules! write_scalar {
 }
 
 pub trait AsyncWriteExt: AsyncWrite {
+    /// Write the entire contents of a buffer into this writer.
+    async fn write_all<T: IoBuf>(&mut self, mut buffer: T) -> BufResult<usize, T> {
+        let buf_len = buffer.buf_len();
+        let mut total_written = 0;
+        while total_written < buf_len {
+            let written;
+            (written, buffer) =
+                buf_try!(self.write(buffer.slice(total_written..)).await.into_inner());
+            total_written += written;
+        }
+        BufResult(Ok(total_written), buffer)
+    }
+
     write_scalar!(u8, to_be_bytes, to_le_bytes);
     write_scalar!(u16, to_be_bytes, to_le_bytes);
     write_scalar!(u32, to_be_bytes, to_le_bytes);
